Move queue utilities
